pyproject.toml
[project]
name = "ainalyst"
version = "0.1.0"
description = "A scratchpad to experiment with a few popular ai packages"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.2",
    "dspy>=2.6.24",
    "ipython>=9.3.0",
    "langchain>=0.3.25",
    "langgraph>=0.4.7",
    "lightrag-hku[api]>=1.3.8",
    "litellm>=1.72.0",
    "lmstudio>=1.3.1",
    "mcp[cli]>=1.9.2",
    "mlx>=0.26.1",
    "mlx-lm>=0.24.1",
    "nano-vectordb>=0.0.4.3",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
]
